// src/typings/api/c2c/advertise/post-advertise.d.ts
export interface AdvertOrderHistoryResp {
    /** 总数 */
    total: number;
    pages: number;
    pageSize: number;
    pageNum: number;
    list: IAdvertOrderHistoryList[];
}

export interface IAdvertOrderHistoryList {
    /** 总价 */
    totalPrice: number;
    /** B 端商户 ID */
    businessId: string;
    /** 状态（CREATED 已创建 (初始值) WAS_PAYED 已付款 WAS_COLLECTED 已收款 WAS_TRANSFER_COIN 已转币 WAS_RECEIVE_COIN 已收币 (已完成) WAS_CANCEL 已取消 NOT_CANCEL__APPEALING 非取消、申诉中 CANCEL__APPEALING 取消、申诉中 NOT_CANCEL__APPEAL_FINISH 非取消、申诉完成 CANCEL__APPEAL_FINISH 取消、申诉完成） */
    statusCd: string;
    /** 广告 ID */
    advertId: string;
    /** 商家用户 ID */
    merchantUid: string;
    /** 用户 ID */
    uid: string;
    /** 数量 */
    number: number;
    /** 交易完成后，链上提币冻结结束时间 */
    freezeEndTime?: any;
    /** 主链 */
    mainchainAddrId?: any;
    /** 支付方式 ID */
    paymentId?: any;
    /** 订单 ID */
    id: string;
    /** 交易区名称 */
    currencyEnName: string;
    /** 币种符号 */
    symbol: string;
    /** 币种全称 */
    coinFullName: string;
    /** 币种名称 */
    coinName: string;
    /** 下单时间 */
    createdTime: string;
    /** 类型 (NUMBER TOTAL_PRICE) */
    typeCd: string;
    /** 广告主用户姓名 */
    advertUserName: string;
    /** 是否解冻过，1 解冻过，2 未解冻过 */
    unfreeze?: number;
    /** 支付（收款）结束时间 */
    expireTime?: any;
    /** 买卖角色 */
    buyAndSellRole: string;
    /** 是否已胜诉 */
    isAppealWinner: boolean;
    /** 买方 - 用户姓名*/
    buyerUserName: string
    /** 单价 */
    price: number
    /** 买方 - 商家昵称 */
    buyerMerNickName?: string
    /** 卖方 - 商家昵称 */
    sellerMerNickName?: string
    /** 买方 - 成单率 */
    buyerCompletedOrderRate?: string
    /** 卖方 - 成单率 */
    sellerCompletedOrderRate?: string
    /** 卖方 - 用户姓名 */
    sellerUserName: string
    /** 买方 - 是否商家 */
    buyerIsMerchant?: boolean
    /** 卖方 - 是否商家 */
    sellerIsMerchant?: boolean
    /** 方向 (BUY 买 SELL 卖) */
    directCd: string
    /** 交易类型 (INSIDE 站内 OUTSIDE 站外)*/
    dealTypeCd: string
    /** 买家用户 ID */
    buyerUid: string
    /** 卖家用户 ID */
    sellerUid: string
    /** 链路名称 */
    mainChainName: string
}